+++ apps/gotask_api/src/domain/model/user/user.ts
@@ -1,18 +1,7 @@
-import { Document, Schema, model, Types } from "mongoose";
+import { Schema, model, Types } from "mongoose";
 import { v4 as uuidv4 } from "uuid";
 import bcrypt from "bcrypt";
-
-// User interface
-export interface IUser extends Document {
-  id: string;
-  name: string;
-  password: string;
-  user_id: string; // email
-  status: boolean;
-  roleId: Types.ObjectId; // renamed from "role"
-  organization?: Types.ObjectId;
-  projects?: Types.ObjectId[];
-}
+import { IUser } from "../../interface/user/userInterface";
 
 // User schema
 const UserSchema = new Schema<IUser>(
@@ -22,8 +11,6 @@ const UserSchema = new Schema<IUser>(
     password: { type: String, required: true },
     user_id: { type: String, required: true, unique: true },
     status: { type: Boolean, default: true },
-
-    // Reference to Role (renamed to roleId)
     roleId: { type: Schema.Types.ObjectId, ref: "Role", required: true },
 
     organization: {
