+++ apps/gotask_api/src/modules/user/userController.ts
@@ -172,4 +172,5 @@ class UserController extends BaseController {
   }
 }
 
+
 export default UserController;
